fix(test): fix broken unit test
